requirements.txt
fastapi
uvicorn[standard]
requests
httpx
beautifulsoup4
python-dotenv